Preserve permissions and ownership for files stored as strings
Files that are stored as strings bypass the permissions code added
in commit 74c1476 since they are added to the archive via
add_string().

Add a check to see if a destination exists in the host file system.
If it does propagate the permissions to the stored string's tar
record.