Rename command output collection interfaces
The method names for collecting external output are ugly and
awkward to type. Rename them to be more consistent with the file
collection API and easier to remember and type.

This change touches virtually every plugin class but is a very
simple search/replace (plugin changes were done with sed).

Fixes Issue #92.