Make PATH policy controlled
Make the PATH environment variable subject to Policy class
control. Policies may set PATH to an appropriately formatted
string for the platform. This will be exported to the environment
and used as the search path for binaries that sos modules execute.

Following this change all absolute paths can be removed from
plug-in code simplifying maintenance on multiple distributions
and distribution versions (e.g. Red Hat UsrMove and various
updates that relocate binaries such as lsusb).

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>