Remove PATH workarounds from hardware plug-in
The hardware plug-in contains code to check for the existence
of binaries before running them. This won't work if plug-ins are
providing non-absolute paths for executables and is redundant in
any case since the shell-out wrappers already do this and log
unavailable binaries. Also removes the workaround for lsusb
changing path in recent versions.

Signed-of-by: Bryn M. Reeves <bmr@redhat.com>