Remove bogus path check in sar module
The sar module tries to check for the existence of a binary at
some exact path. Since all call outs now go through PATH this no
longer works leading to a bogus error level log message.

Remove this check and also move the accompanying check on
/var/lib/sa into a separate check_enabled method - without this
path being present there is nothing for the plug-in to collect.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>