Merge initrd and bootloader plug-ins into a single boot module
The existing initrd plug-in is default-disabled so it rarely
collects anything useful. The command used to extract the init
script from the initramfs image is also broken for modern systems
as the pattern ("initrd*") does not match and the file /init on
systems using systemd is actually a symbolic link to the systemd
binary.

Replace the collection of this file with a listing of the current
initramfs image (via lsinitrd). Optionally collect a listing from
each non-kdmump image present in /boot.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>