Add get_cmd_dir() method to Plugin and make plug-ins use it
As reported in Issue #140 plug-ins that need to write to an
arbitrary location in the archive are broken by the in-line tar
archive changes.

Introduce a 'get_tmp_dir()' method to the Archive classes that
must return a writable directory that is included in the archive.

Archives deriving from FileCacheArchive simply retun the root of
the archive tree.

Add a new method to Plugin to use this support, 'get_cmd_dir()',
that will return 'Archive.get_tmp_dir() + 'sos_commands'
+ self.name()'.

Fixes problems in lvm2, cloudforms, satellite and rhui.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>