Normalize temporary directory path
The archive and temporary file handling classes expect to operate
on absolute paths. If a user specifies a relative path via the
command line --tmp-dir switch it must be normalized before being
passed to these classes to provide correct behaviour. Failing to
do this causes a variety of errors: the final archive cannot be
created since the path is constructed incorrectly and temporary
files are not cleaned up since they are not at the "expected"
location.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>