Fix command output substitution exception
If a comand has a substitution registered via do_cmd_output_sub()
but no data was collected (e.g. command not found) the postproc
code will throw an exception as the return value ('replacements')
is never assigned.

Initialise replacements to None before scanning the list of run
commands and return this if no substitutions were made.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>