[plugin] log directory contents added via Plugin.copy_dir()
We're currently very verbose about the copyspecs and path lists
we process during setup() and collect() but right now we do not
log the same level of information for artefacts that are included
in the archive as a consequence of Plugin.do_copy_path() calling
copy_dir() on the result of copyspec expansion.

Log these explicitly and promote logging of copyspec expansion
to match that used for command logging.

This makes us very chatty at high verbosity but allows a clean
view of exactly what the tree walking code is doing (which my
sanity desperately needs).

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>