[openstack-*,ubuntu_support_status] fix plugin_name
We previously fixed an issue where plugin names needed to have underscores
in the actual filename otherwise they wouldn't be resolved. The `plugin_name`
attribute was never updated to reflect that. This change fixes to match
that of the filename so that when calling something like

sosreport -o ubuntu_support_status

Will work rather than fail with unable to find plugin errors.

Fixes #379

Signed-off-by: Adam Stokes <adam.stokes@ubuntu.com>