[sosreport] re-order logging shutdown
Currently we remove all log handlers and then add the log files
to the archive. This can cause messages to leak to the console
when running in --quiet mode:

  # sosreport -o lvm2 --batch --build --debug --quiet
  No handlers could be found for logger "sos"

This happens because the Archive classes output to the 'sos'
logger; removing the handlers before archiving the logs causes
a default basicConfig to be used and generates the above message.

Add the files to the archive and then remove the handlers.

Fixes #393.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>