[mysql] fix command-line dbpass handling
The logic for chosing between a command-line and environment
variable passed password in the MySQL plugin was incorrect. This
prevents a database dump from being collected when a password is
given on the command line.

Altough use of the command line to pass authentication tokens
is discouraged we should ensure the case works.

Fixes #456.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>