[apparmor] Capture AppArmor profiles and status
Captures /etc/apparmor.d for profiles

Excludes /cache - because it's not config.  Excludes libvirt/libvirt
because it can grow quite large and the TEMPLATE file should usually
be enough. Excludes abstractions because they are usually not
modified and are large.

For both libvirt and abstractions capture an ls just to be sure
permissions are correct.

Captures apparmor_status to get the effective profiles.

Signed-off-by: Bryan Quigley <bryan.quigley@canonical.com>
Signed-off-by: Adam Stokes <adam.stokes@ubuntu.com>