[mysql] Collect log file
When MariaDB is run under Pacemaker an OCF resource agent is used
to start and stop the database. This configuration also overrides
the log paths used for the database. In particular the mysqld.log
is present at '/var/log/mysqld.log'.

Collect both paths as we do not currently detect whether we are
running with an OCF agent or stand-alone.

Fixes: #554

Signed-off-by: Shane Bradley <sbradley@redhat.com>
Signed-off-by: Bryn M. Reeves <bmr@redhat.com>