[sosreport] add --chroot option
Add a --chroot option to sosreport to control command chrooting.

The option takes one of three values:

  * auto   - Allow callers of the API to control chroot behaviour
  * always - Always chroot external commands to --sysroot
  * never  - Never chroot external commands

This is a fairly low-level option and may not be exposed to the
user in a final release; for now it will allow tests in container
environments to control the chrooting behaviour used for a run.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>