[plugin] add tmp_in_sysroot() method
Add a method that plugins can test to determine whether the
archive's temporary directory is inside sysroot. This is always
true when sysroot is '/'. When sysroot is a subdirectory of root
the temporary directory may be inaccessible from the chroot
namespace. Plugins can test this method to determine where to
write output.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>