[networking] gather per-namespace ip and ethtool data
As more applications are beginning to use namespaces, notably OpenStack
and container-based platforms, capturing namespace data is becoming
essential.

This patch alters the networking plugin so that it is namespace aware,
collecting commands such as:

    $ ip address show
    $ ip route
    $ iptables-save

For each individual namespace, along with per-device ethtool
information for any devices that exist in said namespaces.

Closes #642

Signed-off-by: Robb Manes <rmanes@redhat.com>
Signed-off-by: Adam Stokes <adam.stokes@ubuntu.com>