[sosreport] fix command-line report defaults
opts.report is True regardless of using --no-report option or not. That
means the test "if not self.opts.report:" to decide generation of
reports is evaluated to False every time and no report is generated any
time.

This commit renames opts.report and SoSOptions._report to noreport (to
get rid of the variable name confusion) and updates default value
accordingly.

Resolves #625
Closes #644

Signed-off-by: Pavel Moravec <pmoravec@redhat.com>
Signed-off-by: Adam Stokes <adam.stokes@ubuntu.com>