[networking] disable netstat output truncation
By default netstat will truncate addresses that are wider than
the current column width. Disable this to ensure that all IPv6
addresses can be captured in full.

On Red Hat a non-upstream option (-T) was added to versions
prior to 2.0 - handle these version checks in the RedHatNetworking
class and set the appropriate option for each.

Fixes #734.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>