[sosreport] fix '0 plugins' output in profile listings
When sosreport prints a profile listing as a result of the user
selecting a non-existent profile we currently output a profile
list (no plugin names) that ends with the standard summary of
profiles and plugins:

    # sosreport -vv --batch --profile qux
    set sysroot to '/' (default)

    sosreport (version 3.2)

    Unknown or inactive profile(s) provided: qux
    The following profiles are available:

     boot
     cluster
     [...]
     system
     virt
     webserver

     28 profiles, 0 plugins

This is less than useful and confusing for users (who shouldn't
have to care about how we manage this stuff internally).

Change this to only include the ", N plugins" part of the text
if we have actually loaded some valid plugins for the active
profile and factor out common display code in list_plugins and
list_profiles.

Fixes: #733.

Signed-off-by: Bryn M. Reeves <bmr@redhat.com>